Pass the git commit message via a file
Currently we pass it via the command line, which causes a bunch of
grief. Namely: various characters need to be escaped since they will
conflict with the system's restrictions.

Avoid all that and use the --file instead of --message

As a follow-up the no longer required sanitization will be removed.

Issue: https://github.com/hbons/SparkleShare/issues/1788
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>